requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

APScheduler~=3.10.1
boto3 == 1.29.1
loguru == 0.6.0
python_dotenv == 0.20.0
pytz~=2023.3
requests == 2.27.1
tencentcloud_sdk_python_cloudaudit == 3.0.1022
DateTime~=4.3
python-dotenv~=0.20.0
ipaddress~=1.0.23